Extract get_date_range_from_week into class and test it
