+++ core_weekly/core_weekly.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 from .report import Report
 from .template import Template
 from .parser import Parser
+from .date_util import DateUtil
 from pathlib import Path
 import datetime
 import json
@@ -28,7 +29,7 @@ class CoreWeekly():
 
         self.date_range = args.date
         if args.week is not None:
-            self.date_range = self.get_date_range_from_week(args.week, args.year)
+            self.date_range = DateUtil().get_date_range_from_week(args.week, args.year)
 
         if self.date_range:
             self.initialize_time_parameters(self.date_range)
@@ -38,33 +39,6 @@ class CoreWeekly():
 
         self.is_debug = args.debug
         self.directory = Path(__file__).resolve().parents[1] / 'var'
-
-    def get_date_range_from_week(self, week, year):
-        """Get data range from week number
-
-        :param week: Week number
-        :type week: str
-        :param year: Year
-        :type year: str
-        :returns: A date range
-        :rtype: str
-        """
-        if year is None:
-            year = datetime.datetime.now().year
-
-        self.year = year
-        self.week = week
-
-        first_day = datetime.datetime.strptime(f'{year}-W{int(week)-1}-1', "%Y-W%W-%w").date()
-        last_day = first_day + datetime.timedelta(days=6.9)
-
-        self.first_day_number = first_day.strftime('%-d')
-        self.last_day_number = last_day.strftime('%-d')
-
-        self.month = datetime.datetime.strptime(f'{year}-W{int(week)-1}-1', "%Y-W%W-%w").date().strftime('%B')
-
-        return str(first_day) + '..' + str(last_day)
-
 
     def initialize_time_parameters(self, date_range):
